// web/lib/api/generated/model/receiptSupplierName.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface ReceiptSupplierName {
  name: string;
  receipts: number;
  /** @nullable */
  last_date: string | null;
  /** @nullable */
  match_id: string | null;
  match_name: string;
}